Remove all NAs before BRT
This is a sensitive change.

Previously, NAs in the response variable were removed but NAs in the environment data were kept. At locations were some data was available but not all the BRT fitting was still done, with the available data only. This might not be too much of a problem but might still lead to unexpected results. Similarly, during the prediction, when data was not available, it predicted based on the available data only. This is more problematic because if only one variable is present, prediction will be made based on this one only and might not be consistent with what is around.

Now NAs are removed everywhere. This will lead to more "holes" of course.