Better implementation of weights in BRT
It needs to be binned on the environmental data grid, not on the prediction grid.

Need to attach the weights to the data, to make sure they are carried on when this data is modified (when environment data is associated for example). As a side benefit, they are also stored in the resulting object.

Reflect the changes in the documentation.